rag_backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
chromadb==0.5.4
pydantic==2.8.2
huggingface-hub==0.16.4
sentence-transformers==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.9.1 # Version récente et stable
